Rename Relation 'key' to 'field'
This is a little beyond cosmetic, the term 'key' is a little overloaded
in Cassandra because it refers to things that should be unique (such as
primary keys and clustering keys).

Across the rest of the codebase, we use the the phrase 'field' to
represent a logical column. It makes sense to move away from 'key' and
make this consistent, especially as in future we want to expose more of
this to the outside world.