Use allocated zero values of our struct
We had some pretty naive code that would generate the different ptrs
corresponding to the different fields in the struct, allocate the struct
(which would also allocate a bunch of zero values) and then immediately
overwrite them with the new values.

Instead, we can be much smarter and just do a single round of
allocation, when we allocate the struct, Go already gives us the zero
values so we can use them since the struct fields we want to set are
reserved just for us to do assignment!